I created the files HeaderUsers and I deleted Logup becouse it was repeat and i deleted some files thire was not necessary in the such as reportWebVitals.js, setupTests.js, App.test.js, index.css,and  robots.txt
